leopold: Deduplicate acuation point adjustment code
