+++ keyboards/leopold/fc980c/fc980c.c
@@ -23,38 +23,4 @@ void matrix_init_kb(void) {
 
     matrix_init_user();
 }
-
-void actuation_point_up(void) {
-    // write RDAC register: lower value makes actuation point shallow
-    uint8_t rdac = ad5258_read_rdac();
-    if (rdac == 0) {
-        ad5258_write_rdac(0);
-    } else {
-        ad5258_write_rdac(rdac - 1);
-    }
-}
-
-void actuation_point_down(void) {
-    // write RDAC register: higher value makes actuation point deep
-    uint8_t rdac = ad5258_read_rdac();
-    if (rdac == 63) {
-        ad5258_write_rdac(63);
-    } else {
-        ad5258_write_rdac(rdac + 1);
-    }
-}
-
-void adjust_actuation_point(int offset) {
-    ad5258_init();
-    uint8_t rdac = ad5258_read_eeprom() + offset;
-    if (rdac > 63) { // protects from under and overflows
-        if (offset > 0) {
-            ad5258_write_rdac(63);
-        } else {
-            ad5258_write_rdac(0);
-        }
-    } else {
-        ad5258_write_rdac(rdac);
-    }
-}
 #endif
